refactor(Refactor on validator):
Refactor on some functions. Moving the functions inside the validator in a utils file.
