+++ tsconfig.json
@@ -13,7 +13,7 @@
     "noUnusedLocals": true,
     "resolveJsonModule": true,
     "typeRoots" : [
-      "../node_modules/@types"
+      "node_modules/@types"
     ],
   },
   "exclude": [
